+++ lib/ui/settings/debug/debug_settings.dart
@@ -2,6 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:nesd/ui/settings/debug/cartridge_switch.dart';
 import 'package:nesd/ui/settings/debug/debug_overlay_switch.dart';
 import 'package:nesd/ui/settings/debug/debug_tile_switch.dart';
+import 'package:nesd/ui/settings/debug/debugger_switch.dart';
 import 'package:nesd/ui/settings/settings_tab.dart';
 
 class DebugSettings extends StatelessWidget {
@@ -16,6 +17,7 @@ class DebugSettings extends StatelessWidget {
           DebugTileSwitch(),
           CartridgeSwitch(),
           DebugOverlaySwitch(),
+          DebuggerSwitch(),
         ],
       ),
     );