+++ lib/ui/settings/debug/debugger_switch.dart
@@ -1,0 +1,24 @@
+import 'package:flutter/material.dart';
+import 'package:flutter_riverpod/flutter_riverpod.dart';
+import 'package:nesd/ui/common/focus_on_hover.dart';
+import 'package:nesd/ui/common/settings_tile.dart';
+import 'package:nesd/ui/settings/settings.dart';
+
+class DebuggerSwitch extends ConsumerWidget {
+  const DebuggerSwitch({super.key});
+
+  @override
+  Widget build(BuildContext context, WidgetRef ref) {
+    final setting =
+        ref.watch(settingsControllerProvider.select((s) => s.showDebugger));
+    final controller = ref.read(settingsControllerProvider.notifier);
+
+    return FocusOnHover(
+      child: SwitchSettingsTile(
+        title: const Text('Show debugger'),
+        value: setting,
+        onChanged: (value) => controller.showDebugger = value,
+      ),
+    );
+  }
+}