+++ lib/ui/settings/settings.freezed.dart
@@ -26,6 +26,7 @@ mixin _$Settings {
   bool get showTiles => throw _privateConstructorUsedError;
   bool get showCartridgeInfo => throw _privateConstructorUsedError;
   bool get showDebugOverlay => throw _privateConstructorUsedError;
+  bool get showDebugger => throw _privateConstructorUsedError;
   Scaling get scaling => throw _privateConstructorUsedError;
   bool get autoSave => throw _privateConstructorUsedError;
   int? get autoSaveInterval => throw _privateConstructorUsedError;
@@ -64,6 +65,7 @@ abstract class $SettingsCopyWith<$Res> {
       bool showTiles,
       bool showCartridgeInfo,
       bool showDebugOverlay,
+      bool showDebugger,
       Scaling scaling,
       bool autoSave,
       int? autoSaveInterval,
@@ -99,6 +101,7 @@ class _$SettingsCopyWithImpl<$Res, $Val extends Settings>
     Object? showTiles = null,
     Object? showCartridgeInfo = null,
     Object? showDebugOverlay = null,
+    Object? showDebugger = null,
     Object? scaling = null,
     Object? autoSave = null,
     Object? autoSaveInterval = freezed,
@@ -133,6 +136,10 @@ class _$SettingsCopyWithImpl<$Res, $Val extends Settings>
       showDebugOverlay: null == showDebugOverlay
           ? _value.showDebugOverlay
           : showDebugOverlay // ignore: cast_nullable_to_non_nullable
+              as bool,
+      showDebugger: null == showDebugger
+          ? _value.showDebugger
+          : showDebugger // ignore: cast_nullable_to_non_nullable
               as bool,
       scaling: null == scaling
           ? _value.scaling
@@ -189,6 +196,7 @@ abstract class _$$SettingsImplCopyWith<$Res>
       bool showTiles,
       bool showCartridgeInfo,
       bool showDebugOverlay,
+      bool showDebugger,
       Scaling scaling,
       bool autoSave,
       int? autoSaveInterval,
@@ -222,6 +230,7 @@ class __$$SettingsImplCopyWithImpl<$Res>
     Object? showTiles = null,
     Object? showCartridgeInfo = null,
     Object? showDebugOverlay = null,
+    Object? showDebugger = null,
     Object? scaling = null,
     Object? autoSave = null,
     Object? autoSaveInterval = freezed,
@@ -256,6 +265,10 @@ class __$$SettingsImplCopyWithImpl<$Res>
       showDebugOverlay: null == showDebugOverlay
           ? _value.showDebugOverlay
           : showDebugOverlay // ignore: cast_nullable_to_non_nullable
+              as bool,
+      showDebugger: null == showDebugger
+          ? _value.showDebugger
+          : showDebugger // ignore: cast_nullable_to_non_nullable
               as bool,
       scaling: null == scaling
           ? _value.scaling
@@ -307,6 +320,7 @@ class _$SettingsImpl implements _Settings {
       this.showTiles = false,
       this.showCartridgeInfo = false,
       this.showDebugOverlay = false,
+      this.showDebugger = false,
       this.scaling = Scaling.autoInteger,
       this.autoSave = true,
       this.autoSaveInterval = 1,
@@ -345,6 +359,9 @@ class _$SettingsImpl implements _Settings {
   @override
   @JsonKey()
   final bool showDebugOverlay;
+  @override
+  @JsonKey()
+  final bool showDebugger;
   @override
   @JsonKey()
   final Scaling scaling;
@@ -400,7 +417,7 @@ class _$SettingsImpl implements _Settings {
 
   @override
   String toString() {
-    return 'Settings(volume: $volume, stretch: $stretch, showBorder: $showBorder, showTiles: $showTiles, showCartridgeInfo: $showCartridgeInfo, showDebugOverlay: $showDebugOverlay, scaling: $scaling, autoSave: $autoSave, autoSaveInterval: $autoSaveInterval, bindings: $bindings, lastRomPath: $lastRomPath, recentRomPaths: $recentRomPaths, showTouchControls: $showTouchControls, narrowTouchInputConfig: $narrowTouchInputConfig, wideTouchInputConfig: $wideTouchInputConfig)';
+    return 'Settings(volume: $volume, stretch: $stretch, showBorder: $showBorder, showTiles: $showTiles, showCartridgeInfo: $showCartridgeInfo, showDebugOverlay: $showDebugOverlay, showDebugger: $showDebugger, scaling: $scaling, autoSave: $autoSave, autoSaveInterval: $autoSaveInterval, bindings: $bindings, lastRomPath: $lastRomPath, recentRomPaths: $recentRomPaths, showTouchControls: $showTouchControls, narrowTouchInputConfig: $narrowTouchInputConfig, wideTouchInputConfig: $wideTouchInputConfig)';
   }
 
   @override
@@ -418,6 +435,8 @@ class _$SettingsImpl implements _Settings {
                 other.showCartridgeInfo == showCartridgeInfo) &&
             (identical(other.showDebugOverlay, showDebugOverlay) ||
                 other.showDebugOverlay == showDebugOverlay) &&
+            (identical(other.showDebugger, showDebugger) ||
+                other.showDebugger == showDebugger) &&
             (identical(other.scaling, scaling) || other.scaling == scaling) &&
             (identical(other.autoSave, autoSave) ||
                 other.autoSave == autoSave) &&
@@ -446,6 +465,7 @@ class _$SettingsImpl implements _Settings {
       showTiles,
       showCartridgeInfo,
       showDebugOverlay,
+      showDebugger,
       scaling,
       autoSave,
       autoSaveInterval,
@@ -480,6 +500,7 @@ abstract class _Settings implements Settings {
       final bool showTiles,
       final bool showCartridgeInfo,
       final bool showDebugOverlay,
+      final bool showDebugger,
       final Scaling scaling,
       final bool autoSave,
       final int? autoSaveInterval,
@@ -508,6 +529,8 @@ abstract class _Settings implements Settings {
   bool get showCartridgeInfo;
   @override
   bool get showDebugOverlay;
+  @override
+  bool get showDebugger;
   @override
   Scaling get scaling;
   @override